pyproject.toml
[project]
name = "bluehost-log-parser"
version = "1.5.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Todd S", email = "info@tascs.net" }
]
requires-python = ">=3.13"
dependencies = [
    "sqlalchemy-utils>=0.41.2",
    "sqlalchemy>=2.0.39",
    "python-dateutil>=2.9.0.post0",
    "country-converter>=1.3",
    "ipwhois>=1.3.0",
    "pymysql>=1.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
